refactor: Add real links to title tags
Add links to education title tags. Delete raw URL address for online school and turn work titles to paragraphs.
Add styles to uniform paragraphs and anchor tags titles rendering on the page.
Add a right margin to education schools URLs to prevent the anchor tag taking up the entire row. This prevents seing a link handle on the city names that is instead related to the school names.